requirements.txt
aniso8601==8.1.0
attrs==20.3.0
autopep8==1.5.4
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
flask-restx==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pycodestyle==2.6.0
pycparser==2.20
pyOpenSSL==20.0.0
pyrsistent==0.17.3
pytz==2020.4
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
